requirements.txt
aiogram>=3.10
requests>=2.31
python-dotenv>=1.0
aiohttp>=3.8.0